[fix] Remove check if user is going to leave event socket
If the user is leaving the event. They can change their guest
status to 'NotGoing'. This is done through an API call.

Up until now, if the user was 'NotGoing', they could not
leave the socket room that they were in.

In order to fix this bottle neck, the check for seeing if the user
was part of the event has been removed.

This will not cause any crashes, because if a user leaves a room
they are not part of, it will fail silently.